+++ src/cmdline_parser.py
@@ -253,13 +253,24 @@ def parse_args_userinfo(argv):
 
 
 def parse_args_fake_data(argv):
-    description = "Import faked data for Event programs or Data Sets"
-    usage = "\nExample: dhis2-pk fake-data -s play.dhis2.org/demo -u admin -p district"
+    description = "Import fake data for Event programs or Data Sets"
+    usage = "\nExample: dhis2-pk fake-data -s play.dhis2.org/demo -u admin -p district -i kla3mAPgvCH -n 100"
 
     parser = argparse.ArgumentParser(usage=usage, description=description)
     required, optional = standard_arguments(parser)
-    required.add_argument('-i', dest='uid', required=True, help='UID for program or data set to import fake data for')
-    required.add_argument('-n', dest='amount', required=True, type=int, help='Amount of events or data values to fake')
+    required.add_argument(
+        '-i',
+        dest='uid',
+        required=True,
+        help='UID for program or data set to import fake data for. min: 1 - max: 100000'
+    )
+    required.add_argument(
+        '-n',
+        dest='amount',
+        required=True,
+        type=int,
+        help='Amount of events or dataValueSet templates'
+    )
     args = parser.parse_args(argv)
     return get_password(args)